camera-ai/requirements.txt
# BI Management - Camera AI Requirements
# نظام تحليل الكاميرات بالذكاء الاصطناعي

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Computer Vision
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# YOLO Object Detection
ultralytics==8.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Async
asyncio-throttle==1.0.2

# Image Processing
scikit-image==0.22.0
